requirements.txt
asgiref==3.8.1
boto3==1.37.33
botocore==1.37.33
Django==5.2
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
jmespath==1.0.1
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
